+++ packages/back-end/src/util/mongo.util.ts
@@ -1,10 +1,6 @@
 import type { Document } from "mongodb";
-import { MongoClient, Db } from "mongodb";
 import type { Document as MongooseDocument } from "mongoose";
 import mongoose from "mongoose";
-import { getEffectiveAccountPlan } from "shared/enterprise";
-import { UsageLimits } from "back-end/types/organization";
-import { MONGODB_URI } from "back-end/src/util/secrets";
 
 /**
  * Assists in migrating any field options that MongoDB has changed between major versions 3 and 4.
@@ -161,61 +157,3 @@ export function removeMongooseFields<T>(
 export function getCollection(name: string) {
   return mongoose.connection.db.collection(name);
 }
-
-type OrganizationUsage = {
-  limits: {
-    requests: number | "unlimited";
-    bandwidth: number | "unlimited";
-  };
-};
-
-const backendLimits = ({
-  limits: { requests: cdnRequests, bandwidth: cdnBandwidth },
-}: OrganizationUsage) => ({ cdnRequests, cdnBandwidth });
-
-let cachedClient: MongoClient | undefined;
-let cachedDb: Db | undefined;
-
-export const getMongoUsageDb = async () => {
-  if (!cachedClient) {
-    cachedClient = new MongoClient(MONGODB_URI);
-    await cachedClient.connect();
-  }
-
-  if (!cachedDb) cachedDb = await cachedClient.db("usage");
-
-  return cachedDb;
-};
-
-export const closeMongoUsageDb = async () => {
-  if (!cachedClient) return;
-
-  await cachedClient.close();
-  cachedClient = undefined;
-};
-
-export const getOrgUsageLimits = async (
-  organization: string
-): Promise<UsageLimits> => {
-  const collection = (await getMongoUsageDb()).collection<OrganizationUsage>(
-    "organizations_usage"
-  );
-
-  const storedLimits = await collection.findOne({
-    organization,
-  });
-
-  if (storedLimits) return backendLimits(storedLimits);
-
-  const plan = getEffectiveAccountPlan({ id: organization });
-
-  if (plan === "starter") {
-    return { cdnRequests: 1_000_000, cdnBandwidth: 5_000_000_000 };
-  }
-
-  if (plan === "pro" || plan === "pro_sso") {
-    return { cdnRequests: 2_000_000, cdnBandwidth: 20_000_000_000 };
-  }
-
-  return { cdnRequests: "unlimited", cdnBandwidth: "unlimited" };
-};
